AT91 Advanced Interrupt Controller
Emulation of AIC has 32 input pins corresponding to the input interrupt sources
and two output pins corresponding to the IRQ and FIQ pins on the processor. All
features of the controller with the exception of external interrupt handling
are emulated according to the documentation. This includes emulation of the
debug register, edge and level triggered interrupts, software interrupts and
interrupt nesting.

External interrupt handling is not implemented because it differs between
various AT91 family microcontrollers in the number of externally exposed pins.
The only feature of the external interrupt processing that is not implemented
is inverting the level of input signal. Should the need arise, adding proper
implementation of external interrupt logic shouldn't be hard.

Signed-off-by: Filip Navara <filip.navara@gmail.com>