AT91 Debug Unit
The debug unit is implemented as serial character device hosted on the system
bus. Only the most basic features are emulated to allow capturing the output
and receiving external input. No actual timing is emulated based on the baud
rate register, although the information is passed to the backend driver using
the qemu_chr_ioctl API.

A two-byte FIFO transmission register, channel mode (normal, local feedback,
remote feedback) and in-circuit emulator (ICE) pins are not emulated. Emulating
the first two features would be straightforward. Emulating the ICE
communication is possible, but changes to CPU emulation are required to
accomplish it. The benefits for debugging would probably not be very high since
the built-in QEMU debugger is much more powerful.

Signed-off-by: Filip Navara <filip.navara@gmail.com>