AT91 Ethernet MAC
The Ethernet controller is implemented as a device on the system bus. It uses
DMA to read and write data from the main memory for packet queues.

Most of the controller features are implemented with the following exceptions:
- Statistics and related registers
- Support of VLAN tags
- Support for transmitting jumbo frames
- Recognizing of MAC addresses by a hash function

The emulation is based on the AT91SAM7X specification and also supports
registers present in the older AT91RM9200 controller - notably the TAR
register for sending a single frame by writing a physical memory
address into it.

EMAC also provides management interface for directly talking to the PHY chip.
The emulation of PHY chip is currently hard-coded in the EMAC emulation and
behaves as a DM9161 chip. Only few registers of the PHY chip are implemented.
Nevertheless it's enough to run the Atmel examples and even FreeRTOS demos.

Signed-off-by: Filip Navara <filip.navara@gmail.com>