TCG constant propagation.
Add support for constant propagation to TCG. This has to be paired with the liveness
analysis to remove the dead code. Not all possible operations are covered, but the
most common ones are. This improves the code generation for several ARM instructions,
like MVN (immediate), and it may help other targets as well.

v1 -> v2:
Added profiling code and hopefully fixed for 64-bit TCG targets.
v2 -> v3:
Another attempt at fixing the support for 64-bit TCG targets.

Signed-off-by: Filip Navara <filip.navara@gmail.com>