GPIO matrix keyboard implementation
This patch introduces an emulation of the standard 4x4 matrix keyboard that
could be connected to GPIO controller.

It doesn't emulate any pull-up resistors connected to the wires and depends on
the GPIO controller for their emulation. This requires the GPIO controller to
support the notion of special -1 value on the GPIO pins, which signals that
the wire is currently disconnected.

The emulation of keys is implemented through standard QEMU keyboard emulation
capabilities. An individual key mapping is specified using a device property
"keys", which should be set to array of 16 bytes where each byte corresponds
to scan code of the key.

Signed-off-by: Filip Navara <filip.navara@gmail.com>

Conflicts:

	Makefile.target