GPIO rotary encoder implementation
This patch introduces an emulation of the rotary encoder that could be connected
to GPIO controller.

An internal state is maintained with the values of the output pins. When a key is
pressed the state is modified accordingly and the information is signaled on the
output pins.

Key mappings are specified by the "key-left" and "key-left-alt" properties for a
counter-clockwise direction and "key-right" and "key-right-alt" for a clockwise
direction.

Signed-off-by: Filip Navara <filip.navara@gmail.com>

Conflicts:

	Makefile.target